Move filters from handler file
