+++ src/handlers/events.rs
@@ -1,6 +1,6 @@
 use warp::{http::StatusCode, Filter};
-use serde::{Serialize, Deserialize};
-use crate::models::models::{Account, Debtor, Item};
+use serde::{Serialize};
+use crate::models::models::{Account};
 use std::convert::Infallible;
 use crate::models::{DataManager};
 use crate::models::requests::{Payment, CreateAccount};
@@ -10,80 +10,9 @@ extern crate rand;
 use rand::Rng; 
 use rand::distributions::Alphanumeric;
 
-
 #[derive(Serialize, Debug, Clone)]
 pub struct CustomError{
     pub error: String
-}
-
-fn with_db(db: impl DataManager) -> impl Filter<Extract = (impl DataManager,), Error = Infallible> + Clone {
-    warp::any().map(move || db.clone())
-}
-
-fn json_body() -> impl Filter<Extract = (CreateAccount,), Error = warp::Rejection> + Clone {
-    // When accepting a body, we want a JSON body
-    // (and to reject huge payloads)...
-    warp::body::content_length_limit(1024 * 16).and(warp::body::json())
-}
-
-pub fn events_endpoint(
-    db: impl DataManager,
-) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
-        event_create(db.clone())
-        .or(event_get(db.clone()))
-        .or(event_get_by_user(db.clone()))
-        .or(event_make_payment(db.clone()))
-}
-
-/// POST /events/create with JSON body
-pub fn event_create(
-    db: impl DataManager,
-) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
-    warp::path!("api" / "v1" / "events")
-        .and(warp::post())
-        .and(json_body())
-        .and(with_db(db))
-        .and_then(create_event)
-}
-
-/// GET /events/{events_id}
-pub fn event_get(
-    db: impl DataManager,
-) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
-    warp::path!("api" / "v1" / "events" / String)
-        .and(warp::get())
-        .and(with_db(db))
-        .and_then(event_info)
-}
-
-pub fn event_get_by_user(
-    db: impl DataManager,
-) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
-    warp::path!("api" / "v1" /  "user" / String)
-        .and(warp::get())
-        .and(with_db(db))
-        .and_then(user_events)
-}
-
-/// POST /events/pay with JSON body
-pub fn event_make_payment(
-    db: impl DataManager,
-) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
-    warp::path!("api" / "v1" / "events" / "pay")
-        .and(warp::post())
-        .and(json_body())
-        .and(with_db(db))
-        .and_then(create_event)
-}
-
-pub fn event_update(
-    db: impl DataManager,
-) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
-    warp::path!("api" / "v1" / "events" / String)
-        .and(warp::put())
-        .and(json_body())
-        .and(with_db(db))
-        .and_then(update_event)
 }
 
 pub async fn event_info(id: String, db: impl DataManager) -> Result<impl warp::Reply, Infallible>{
@@ -114,7 +43,7 @@ pub async fn user_events(username: String, db: impl DataManager) -> Result<impl 
         Ok(acc) => Ok(            
             warp::reply::with_status(
                 warp::reply::json(&acc),
-                StatusCode::CREATED
+                StatusCode::OK
             )
         ),
         Err(_e) =>  Ok(
@@ -150,7 +79,7 @@ pub async fn create_event(create: CreateAccount, db: impl DataManager) -> Result
 pub async fn update_event(id: String, acc: CreateAccount, db: impl DataManager) -> Result<impl warp::Reply, Infallible>{
 
     match db.update_account(id, acc.clone()){
-        Ok(_)=> Ok(warp::reply::with_status(warp::reply::json(&acc), StatusCode::OK)),
+        Ok(a)=> Ok(warp::reply::with_status(warp::reply::json(&a), StatusCode::ACCEPTED)),
         Err(e) => Ok(warp::reply::with_status(warp::reply::json(&e), StatusCode::NOT_FOUND)),
     }
 }
